Big commit - Model though working
There are different files/folders in the recommender-system folder:
update-users-marks : Use to update the marks user/user (based on
the nb of messages exchanged)
model-x : User part of the model. Supposed to be in local (user device)
in the alpha version.
mapping: Necessary to create a mapping user_id_raw (real) : user_id
(mathematical id for matrix computations). To call when the user_id_raw
is created.
app.py in the recommender model implements the Y Master-Online part of
the model